[Trivial] Rename specialtx.{cpp|h} to specialtx_validation.{cpp|h}
