+++ src/evo/specialtx_validation.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "evo/specialtx.h"
+#include "evo/specialtx_validation.h"
 
 #include "chain.h"
 #include "coins.h"
