+++ test/lint/lint-circular-dependencies.sh
@@ -64,7 +64,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "qt/pivx/coldstakingwidget -> qt/pivx/tooltipmenu -> qt/pivx/pivxgui -> qt/pivx/coldstakingwidget"
     "qt/pivx/masternodeswidget -> qt/pivx/tooltipmenu -> qt/pivx/pivxgui -> qt/pivx/masternodeswidget"
     "qt/pivx/pivxgui -> qt/pivx/send -> qt/pivx/tooltipmenu -> qt/pivx/pivxgui"
-    "chain -> legacy/stakemodifier -> validation -> evo/specialtx -> chain"
+    "chain -> legacy/stakemodifier -> validation -> evo/specialtx_validation -> chain"
     "chain -> legacy/stakemodifier -> validation -> validationinterface -> chain"
     "chain -> legacy/stakemodifier -> stakeinput -> txdb -> chain"
     "chain -> legacy/stakemodifier -> validation -> checkpoints -> chain"
@@ -73,7 +73,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chain -> legacy/stakemodifier -> validation -> zpivchain -> chain"
     "consensus/tx_verify -> consensus/zerocoin_verify -> spork -> validation -> consensus/tx_verify"
     "evo/deterministicmns -> masternodeman -> net_processing -> validationinterface -> evo/deterministicmns"
-    "evo/deterministicmns -> masternodeman -> validation -> evo/specialtx -> evo/deterministicmns"
+    "evo/deterministicmns -> masternodeman -> validation -> evo/specialtx_validation -> evo/deterministicmns"
     "chain -> legacy/stakemodifier -> stakeinput -> wallet/wallet -> spork -> net_processing -> chain"
 )
